+++ src/interactive/azext_interactive/azclishell/app.py
@@ -36,7 +36,7 @@ from prompt_toolkit.shortcuts import create_eventloop
 
 from . import VERSION
 from .az_completer import AzCompleter
-from .az_lexer import get_az_lexer, ExampleLexer, ToolbarLexer
+from .az_lexer import get_az_lexer, ExampleLexer, ToolbarLexer, ScenarioLexer
 from .configuration import Configuration, SELECT_SYMBOL
 from .frequency_heuristic import DISPLAY_TIME, frequency_heuristic
 from .gather_commands import add_new_lines, GatherCommands
@@ -44,6 +44,8 @@ from .key_bindings import InteractiveKeyBindings
 from .layout import LayoutManager
 from .progress import progress_view
 from . import telemetry
+from .recommendation import Recommender, _show_details_for_e2e_scenario, gen_command_in_scenario
+from .scenario_suggest import ScenarioAutoSuggest
 from .threads import LoadCommandTableThread
 from .util import get_window_dim, parse_quotes, get_os_clear_screen_word
 
@@ -66,8 +68,13 @@ def space_toolbar(settings_items, empty_space):
     if len(settings_items) == 1:
         spacing = ''
     else:
-        spacing = empty_space[
-            :int(math.floor((len(empty_space) - counter) / (len(settings_items) - 1)))]
+        # Calculate the length of space between items
+        spacing_len = (len(empty_space) - len(NOTIFICATIONS) - counter) // (len(settings_items) - 1)
+        if spacing_len < 0:
+            # Not display the first item of settings if the space is not enough
+            return space_toolbar(settings_items[1:], empty_space)
+        # Sample spacing string from `empty_space`
+        spacing = empty_space[:spacing_len]
 
     settings = spacing.join(settings_items)
 
@@ -110,7 +117,7 @@ class AzInteractiveShell(object):
         self.param_docs = u''
         self.example_docs = u''
         self.last = None
-        self.last_exit = 0
+        self.last_exit_code = 0
         self.user_feedback = user_feedback
         self.input = input_custom
         self.output = output_custom
@@ -122,6 +129,9 @@ class AzInteractiveShell(object):
         self.intermediate_sleep = intermediate_sleep
         self.final_sleep = final_sleep
         self.command_table_thread = None
+        self.recommender = Recommender(
+            self.cli_ctx, os.path.join(self.config.get_config_dir(), self.config.get_recommend_path()))
+        self.recommender.set_on_prepared_callback(self.redraw_scenario_recommendation_info)
 
         # try to consolidate state information here...
         # Used by key bindings and layout
@@ -143,6 +153,9 @@ class AzInteractiveShell(object):
         if not self.config.has_feedback() and frequency_heuristic(self):
             print("\n\nAny comments or concerns? You can use the \'feedback\' command!" +
                   " We would greatly appreciate it.\n")
+
+        print("\nA new Recommender is added which can make the completion ability more intelligent and provide the scenario completion!\n"
+              "If you don't want to enable this feature, you can use 'az config set interactive.enable_recommender=False' to disable it.\n")
 
         self.cli_ctx.data["az_interactive_active"] = True
         self.run()
@@ -205,6 +218,18 @@ class AzInteractiveShell(object):
             self.lexer = get_az_lexer(command_info)
         self._cli = None
 
+    def redraw_scenario_recommendation_info(self):
+        scenarios = self.recommender.get_scenarios() or []
+        scenarios_rec_info = "Scenario Recommendation: "
+        for idx, s in enumerate(scenarios):
+            idx_display = f'[{idx+1}]'
+            scenario_desc = f'{s["scenario"]}'
+            command_size = f'{len(s["nextCommandSet"])} Commands'
+            scenarios_rec_info += f'\n {idx_display} {scenario_desc} ({command_size})'
+        self.cli.buffers['scenarios'].reset(
+            initial_document=Document(u'{}'.format(scenarios_rec_info)))
+        self.cli.request_redraw()
+
     def _space_examples(self, list_examples, rows, section_value):
         """ makes the example text """
         examples_with_index = []
@@ -239,18 +264,19 @@ class AzInteractiveShell(object):
     def _update_toolbar(self):
         cli = self.cli
         _, cols = get_window_dim()
-        cols = int(cols)
+        # The rightmost column in window doesn't seem to be used
+        cols = int(cols) - 1
 
         empty_space = " " * cols
 
         delta = datetime.datetime.utcnow() - START_TIME
         if self.user_feedback and delta.seconds < DISPLAY_TIME:
             toolbar = [
-                ' Try out the \'feedback\' command',
+                'Try out the \'feedback\' command',
                 'If refreshed disappear in: {}'.format(str(DISPLAY_TIME - delta.seconds))]
         elif self.command_table_thread.is_alive():
             toolbar = [
-                ' Loading...',
+                'Loading...',
                 'Hit [enter] to refresh'
             ]
         else:
@@ -259,6 +285,8 @@ class AzInteractiveShell(object):
         toolbar, empty_space = space_toolbar(toolbar, empty_space)
         cli.buffers['bottom_toolbar'].reset(
             initial_document=Document(u'{}{}{}'.format(NOTIFICATIONS, toolbar, empty_space)))
+        # Reset the cursor pos so that the bottom toolbar doesn't appear offset
+        cli.buffers['bottom_toolbar'].cursor_position = 0
 
     def _toolbar_info(self):
         sub_name = ""
@@ -273,9 +301,10 @@ class AzInteractiveShell(object):
         tool_val = 'Subscription: {}'.format(sub_name) if sub_name else curr_cloud
 
         settings_items = [
-            " [F1]Layout",
+            "[F1]Layout",
             "[F2]Defaults",
             "[F3]Keys",
+            "[Space]Predict",
             "[Ctrl+D]Quit",
             tool_val
         ]
@@ -297,8 +326,11 @@ class AzInteractiveShell(object):
         if not self.completer.complete_command and new_command in self.completer.command_description:
             command = new_command
 
-        # get command/group help
-        if self.completer and command in self.completer.command_description:
+        if not command and self.recommender.enabled:
+            # display hint to promote CLI recommendation when the user doesn't have any input
+            self.description_docs = u'Try [Space] or `next` to get Command Recommendation'
+        elif self.completer and command in self.completer.command_description:
+            # Display the help message of the command when the user has input
             self.description_docs = u'{}'.format(self.completer.command_description[command])
 
         # get parameter help if full command
@@ -335,11 +367,12 @@ class AzInteractiveShell(object):
         except configparser.NoSectionError:
             self.config_default = ""
 
-    def create_application(self, full_layout=True):
+    def create_application(self, full_layout=True, auto_suggest=AutoSuggestFromHistory(),
+                           prompt_prefix='', toolbar_hint=''):
         """ makes the application object and the buffers """
-        layout_manager = LayoutManager(self)
+        layout_manager = LayoutManager(self, prompt_prefix, toolbar_hint)
         if full_layout:
-            layout = layout_manager.create_layout(ExampleLexer, ToolbarLexer)
+            layout = layout_manager.create_layout(ExampleLexer, ToolbarLexer, ScenarioLexer)
         else:
             layout = layout_manager.create_tutorial_layout()
 
@@ -352,12 +385,13 @@ class AzInteractiveShell(object):
             'example_line': Buffer(is_multiline=True),
             'default_values': Buffer(),
             'symbols': Buffer(),
-            'progress': Buffer(is_multiline=False)
+            'progress': Buffer(is_multiline=False),
+            'scenarios': Buffer(is_multiline=True, read_only=True),
         }
 
         writing_buffer = Buffer(
             history=self.history,
-            auto_suggest=AutoSuggestFromHistory(),
+            auto_suggest=auto_suggest,
             enable_history_search=True,
             completer=self.completer,
             complete_while_typing=Always()
@@ -398,17 +432,19 @@ class AzInteractiveShell(object):
 
     def handle_example(self, text, continue_flag):
         """ parses for the tutorial """
+        # Get the cmd part and index part from input
+        # e.g. webapp create :: 1  => `cmd = 'webapp create'` `selected_option='1'`
         cmd = text.partition(SELECT_SYMBOL['example'])[0].rstrip()
-        num = text.partition(SELECT_SYMBOL['example'])[2].strip()
+        selected_option = text.partition(SELECT_SYMBOL['example'])[2].strip()
         example = ""
         try:
-            num = int(num) - 1
+            selected_option = int(selected_option) - 1
         except ValueError:
             print("An Integer should follow the colon", file=self.output)
             return ""
         if cmd in self.completer.command_examples:
-            if num >= 0 and num < len(self.completer.command_examples[cmd]):
-                example = self.completer.command_examples[cmd][num][1]
+            if 0 <= selected_option < len(self.completer.command_examples[cmd]):
+                example = self.completer.command_examples[cmd][selected_option][1]
                 example = example.replace('\n', '')
             else:
                 print('Invalid example number', file=self.output)
@@ -433,14 +469,19 @@ class AzInteractiveShell(object):
         return self.example_repl(example_no_fill, example, starting_index, continue_flag)
 
     def example_repl(self, text, example, start_index, continue_flag):
-        """ REPL for interactive tutorials """
+        """ REPL(Read-Eval-Print Loop) for interactive tutorials """
         if start_index:
             start_index = start_index + 1
             cmd = ' '.join(text.split()[:start_index])
             example_cli = CommandLineInterface(
                 application=self.create_application(
-                    full_layout=False),
+                    full_layout=False,
+                    prompt_prefix='(tutorial) ',
+                    toolbar_hint='In Tutorial Mode: Press [Enter] after typing each part'
+                ),
                 eventloop=create_eventloop())
+            # Scenario recommendation cannot be enabled in tutorial mode
+            self.completer.enable_scenario_recommender(False)
             example_cli.buffers['example_line'].reset(
                 initial_document=Document(u'{}\n'.format(
                     add_new_lines(example)))
@@ -464,12 +505,94 @@ class AzInteractiveShell(object):
                         start_index += 1
                         cmd += " " + answer.split()[-1] + " " +\
                                u' '.join(text.split()[start_index:start_index + 1])
+            self.completer.enable_scenario_recommender(True)
             example_cli.exit()
             del example_cli
         else:
             cmd = text
 
         return cmd, continue_flag
+
+    def handle_scenario(self, text):
+        """ parses for the scenario recommendation """
+        # Get the index part from input
+        # e.g. :: 1  => `selected_option='1'`
+        selected_option = text.partition(SELECT_SYMBOL['example'])[2].strip()
+        try:
+            selected_option = int(selected_option) - 1
+        except ValueError:
+            print("An Integer should follow the colon", file=self.output)
+            return
+        if 0 <= selected_option < len(self.recommender.get_scenarios() or []):
+            scenario = self.recommender.get_scenarios()[selected_option]
+            self.recommender.feedback_scenario(selected_option, scenario)
+        else:
+            print('Invalid example number', file=self.output)
+            return
+        self.scenario_repl(scenario)
+
+    def scenario_repl(self, scenario):
+        """ REPL(Read-Eval-Print Loop) for interactive scenario execution """
+        auto_suggest = ScenarioAutoSuggest()
+        example_cli = CommandLineInterface(
+            application=self.create_application(
+                full_layout=False,
+                auto_suggest=auto_suggest,
+                prompt_prefix='(scenario) ',
+                toolbar_hint='In Scenario Mode: Press [Enter] to execute commands   [Ctrl+C]Skip  [Ctrl+D]Quit'
+            ),
+            eventloop=create_eventloop())
+        # When users execute the recommended command combination of scenario,
+        # they no longer need the scenario recommendation
+        self.completer.enable_scenario_recommender(False)
+
+        _show_details_for_e2e_scenario(scenario, file=self.output)
+
+        example_cli.buffers['example_line'].reset(
+            initial_document=Document(scenario.get('reason') or scenario.get('scenario') or 'Running a E2E Scenario. ')
+        )
+        quit_scenario = False
+        for nx_cmd, sample in gen_command_in_scenario(scenario, file=self.output):
+            auto_suggest.update(sample)
+            retry = True
+            while retry:
+                # reset and put the command in write buffer
+                example_cli.buffers[DEFAULT_BUFFER].reset(
+                    initial_document=Document(
+                        u'{}'.format(nx_cmd['command']),
+                        cursor_position=len(nx_cmd['command'])))
+                example_cli.request_redraw()
+                try:
+                    # wait for user's input
+                    document = example_cli.run()
+                except (KeyboardInterrupt, ValueError):
+                    # CTRL C
+                    break
+                if not document:
+                    # CTRL D
+                    quit_scenario = True
+                    break
+                if not document.text:
+                    continue
+                cmd = document.text
+                self.history.append(cmd)
+                # Prefetch the next recommendation using current executing command
+                self.recommender.update_executing(cmd, feedback=False)
+                telemetry.start()
+                self.cli_execute(cmd)
+                if self.last_exit_code:
+                    telemetry.set_failure()
+                else:
+                    retry = False
+                    telemetry.set_success()
+                # Update execution result of previous command, fetch recommendation if command failed
+                self.recommender.update_exec_result(self.last_exit_code, telemetry.get_error_info()['result_summary'])
+                telemetry.flush()
+            if quit_scenario:
+                break
+        self.completer.enable_scenario_recommender(True)
+        example_cli.exit()
+        del example_cli
 
     # pylint: disable=too-many-statements
     def _special_cases(self, cmd, outside):
@@ -498,9 +621,9 @@ class AzInteractiveShell(object):
             telemetry.track_outside_gesture()
 
         elif cmd_stripped[0] == SELECT_SYMBOL['exit_code']:
-            meaning = "Success" if self.last_exit == 0 else "Failure"
+            meaning = "Success" if self.last_exit_code == 0 else "Failure"
 
-            print(meaning + ": " + str(self.last_exit), file=self.output)
+            print(meaning + ": " + str(self.last_exit_code), file=self.output)
             continue_flag = True
             telemetry.track_exit_code_gesture()
         elif SELECT_SYMBOL['query'] in cmd_stripped and self.last and self.last.result:
@@ -514,6 +637,9 @@ class AzInteractiveShell(object):
                 self.cli_ctx.show_version()
             except SystemExit:
                 pass
+        elif cmd.startswith(SELECT_SYMBOL['example']):
+            self.handle_scenario(cmd)
+            continue_flag = True
         elif SELECT_SYMBOL['example'] in cmd:
             cmd, continue_flag = self.handle_example(cmd, continue_flag)
             telemetry.track_ran_tutorial()
@@ -662,7 +788,7 @@ class AzInteractiveShell(object):
             else:
                 result = invocation.execute(args)
 
-            self.last_exit = 0
+            self.last_exit_code = 0
             if result and result.result is not None:
                 if self.output:
                     self.output.write(result)
@@ -673,9 +799,9 @@ class AzInteractiveShell(object):
                     self.last = result
 
         except Exception as ex:  # pylint: disable=broad-except
-            self.last_exit = handle_exception(ex)
+            self.last_exit_code = handle_exception(ex)
         except SystemExit as ex:
-            self.last_exit = int(ex.code)
+            self.last_exit_code = int(ex.code)
 
     def progress_patch(self, *args, **kwargs):
         """ forces to use the Shell Progress """
@@ -700,6 +826,8 @@ class AzInteractiveShell(object):
         telemetry.flush()
         while True:
             try:
+                # Reset the user input analysis state in completer to clear the hint in command description section
+                self.completer.reset()
                 document = self.cli.run(reset_current_buffer=True)
                 text = document.text
                 if not text:
@@ -732,10 +860,14 @@ class AzInteractiveShell(object):
                     subprocess.Popen(cmd, shell=True).communicate()
                 else:
                     telemetry.start()
+                    # Prefetch the next recommendation using current executing command
+                    self.recommender.update_executing(cmd)
                     self.cli_execute(cmd)
-                    if self.last_exit and self.last_exit != 0:
+                    if self.last_exit_code:
                         telemetry.set_failure()
                     else:
                         telemetry.set_success()
+                    # Update execution result of previous command, fetch recommendation if command failed
+                    self.recommender.update_exec_result(self.last_exit_code, telemetry.get_error_info()['result_summary'])
                     telemetry.flush()
         telemetry.conclude()